Refactor rotation code to use a standardised naming convention with new code
