+++ plugins/UsersManager/Emails/AuthTokenExpirationWarningNotificationEmail.php
@@ -12,15 +12,15 @@ namespace Piwik\Plugins\UsersManager\Emails;
 use Piwik\Config;
 use Piwik\Mail;
 use Piwik\Piwik;
-use Piwik\Plugins\UsersManager\TokenNotifications\TokenNotification;
+use Piwik\Plugins\UsersManager\TokenNotifications\AuthTokenEmailExpirationWarningNotification;
 use Piwik\SettingsPiwik;
 use Piwik\Url;
 use Piwik\View;
 
-class AuthTokenNotificationEmail extends Mail
+class AuthTokenExpirationWarningNotificationEmail extends Mail
 {
     /**
-     * @var TokenNotification
+     * @var AuthTokenEmailExpirationWarningNotification
      */
     private $notification;
 
@@ -30,8 +30,11 @@ class AuthTokenNotificationEmail extends Mail
     /** @var array */
     private $emailData;
 
-    public function __construct(TokenNotification $notification, string $recipient, array $emailData)
-    {
+    public function __construct(
+        AuthTokenExpirationWarningEmailNotification $notification,
+        string $recipient,
+        array $emailData
+    ) {
         parent::__construct();
 
         $this->notification = $notification;
@@ -51,11 +54,10 @@ class AuthTokenNotificationEmail extends Mail
         $this->setWrappedHtmlBody($this->getDefaultBodyView());
     }
 
-    private function getRotationPeriodPretty(): string
+    private function getExpirationWarningPeriodPretty(): string
     {
-        $rotationPeriodDays = Config::getInstance()->General['auth_token_rotation_notification_days'];
-
-        return $rotationPeriodDays . ' ' . Piwik::translate('Intl_PeriodDay' . ($rotationPeriodDays === 1 ? '' : 's'));
+        $expirationPeriodDays = Config::getInstance()->General['auth_token_expiration_notification_days'];
+        return $expirationPeriodDays . ' ' . Piwik::translate('Intl_PeriodDay' . ($expirationPeriodDays === 1 ? '' : 's'));
     }
 
     protected function getManageAuthTokensLink(): string
@@ -67,12 +69,17 @@ class AuthTokenNotificationEmail extends Mail
     }
     protected function getDefaultSubject(): string
     {
-        return Piwik::translate('UsersManager_AuthTokenNotificationEmailSubject');
+        return Piwik::translate(
+            'UsersManager_AuthTokenExpirationWarningEmailSubject',
+            [
+                $this->getExpirationWarningPeriodPretty()
+            ]
+        );
     }
 
     protected function getDefaultBodyText(): string
     {
-        $view = new View('@UsersManager/_authTokenNotificationTextEmail.twig');
+        $view = new View('@UsersManager/_authTokenExpirationWarningNotificationTextEmail.twig');
         $view->setContentType('text/plain');
 
         $this->assignCommonParameters($view);
@@ -82,7 +89,7 @@ class AuthTokenNotificationEmail extends Mail
 
     protected function getDefaultBodyView(): View
     {
-        $view = new View('@UsersManager/_authTokenNotificationHtmlEmail.twig');
+        $view = new View('@UsersManager/_authTokenExpirationWarningNotificationHtmlEmail.twig');
 
         $this->assignCommonParameters($view);
 
@@ -92,9 +99,9 @@ class AuthTokenNotificationEmail extends Mail
     protected function assignCommonParameters(View $view): void
     {
         $view->tokenName = $this->notification->getTokenName();
-        $view->tokenCreationDate = $this->notification->getTokenCreationDate();
+        $view->tokenExpirationDate = $this->notification->getTokenExpirationDate();
 
-        $view->rotationPeriod = $this->getRotationPeriodPretty();
+        $view->expirationWarningPeriod = $this->getExpirationWarningPeriodPretty();
         $view->manageAuthTokensLink = $this->getManageAuthTokensLink();
 
         foreach ($this->emailData as $item => $value) {
