+++ plugins/UsersManager/templates/_authTokenExpirationWarningNotificationTextEmail.twig
