+++ plugins/UsersManager/templates/_authTokenRotationNotificationTextEmail.twig
